websocket-listener/requirements.txt
alpaca-py==0.16.0
redis==5.0.3
nest_asyncio